Input (#8)
- Input component, with supporting Hint, Label and ErrorMessage components.
- Use template tests from `govuk-frontend` via `govuk-frontend-template-spec`

Note: At the moment, Jest snapshots are manually copied across. When Jest 0.24 comes out there will be a feature to allow the snapshot test location to be configured (https://github.com/facebook/jest/pull/6143), so will keep the manual process for now, with the aim to load them in from node_modules when 0.24 is out.